refactor(acp): extract permission response attempt owner
